Namespaced presenters and moved to lib
