--- app/presenters/view_all_filter.rb
@@ -1,12 +1,0 @@
-class ViewAllFilter < DashboardFilter
-  include Rails.application.routes.url_helpers
-
-  def active?
-    @params[:controller] == 'dashboard' && @params[:action] == 'index' &&
-      @params[:qstatus].blank?
-  end
-
-  def path
-    dashboard_path
-  end
-end
