+++ lib/presenters.rb
@@ -1,0 +1,2 @@
+module Presenters
+end
