+++ lib/presenters/iww_filter.rb
@@ -1,0 +1,13 @@
+module Presenters
+  class IwwFilter < DashboardFilter
+    include Rails.application.routes.url_helpers
+
+    def active?
+      @params[:controller] == 'dashboard' && @params[:action] == 'i_will_write'
+    end
+
+    def path
+      dashboard_iww_path
+    end
+  end
+end
