+++ lib/presenters/report/row.rb
@@ -1,0 +1,16 @@
+module Presenters
+  class Report
+    class Row
+      attr_reader :cells, :state
+
+      def initialize(state, cells)
+        @state = state
+        @cells = cells
+      end
+
+      def label
+        PQState.state_label(@state)
+      end
+    end
+  end
+end
